Changed Float32 to Float64

+++ src/Common/ThreadProfileEvents.cpp
@@ -560,7 +560,7 @@ void PerfEventsCounters::finalizeProfileEvents(ProfileEvents::Counters & profile
         // deltas from old values.
         const auto enabled = current_value.time_enabled - previous_value.time_enabled;
         const auto running = current_value.time_running - previous_value.time_running;
-        const auto scaled_value = (current_value.value - previous_value.value) * enabled / std::max(1.f, float(running));
+        const auto scaled_value = static_cast<Float64>(current_value.value - previous_value.value) * static_cast<Float64>(enabled) / std::max(1., static_cast<Float64>(running));
 
         UInt64 delta = 0;
